api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
python-dotenv==1.0.0
redis==5.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
alembic==1.12.1
strawberry-graphql==0.215.0
aioredis==2.0.1 